+++ typescript.js
@@ -5,6 +5,7 @@ import {
 import getJsdocEslintParser from './src/getJsdocEslintParser.js';
 
 export const parseForESLint = getJsdocEslintParser(typescriptEslintParser, {
+  mode: 'typescript',
   tokens: true,
   comment: true,
   loc: true,